backend/requirements.txt
django
django-cors-headers
opencv-python
spacy
nltk
numpy
Pillow
openai
boto3
djangorestframework
